Remove unneeded file
