+++ message/infrastructure/sql/pubsub_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/ThreeDotsLabs/watermill/message/infrastructure"
 	"github.com/ThreeDotsLabs/watermill/message/infrastructure/sql"
 
+	driver "github.com/go-sql-driver/mysql"
 	"github.com/stretchr/testify/require"
 )
 
@@ -44,6 +45,24 @@ func newPubSub(t *testing.T, db *std_sql.DB, consumerGroup string) message.PubSu
 	require.NoError(t, err)
 
 	return message.NewPubSub(publisher, subscriber)
+}
+
+func newMySQL(t *testing.T) *std_sql.DB {
+	conf := driver.Config{
+		User:                 "root",
+		Passwd:               "",
+		Net:                  "",
+		Addr:                 "localhost",
+		DBName:               "watermill",
+		AllowNativePasswords: true,
+	}
+	db, err := std_sql.Open("mysql", conf.FormatDSN())
+	require.NoError(t, err)
+
+	err = db.Ping()
+	require.NoError(t, err)
+
+	return db
 }
 
 func createPubSubWithConsumerGroup(t *testing.T, consumerGroup string) infrastructure.PubSub {
